[Gate] Rename json_mnger to luos_to_json
