+++ Apps/Gate/convert.h
@@ -8,10 +8,8 @@
 #ifndef CONVERT_H_
 #define CONVERT_H_
 
-#include <json_mnger.h>
-#include "cJSON.h"
-#include "container_structs.h"
 #include "luos.h"
+#include "cJSON.h"
 
 /*
  * Pid
