+++ Apps/Gate/luos_to_json.h
@@ -1,8 +1,7 @@
 
-#ifndef ROUTINGTABLE
-#define ROUTINGTABLE
+#ifndef LUOS_TO_JSON
+#define LUOS_TO_JSON
 
-#include "luos_list.h"
 #include "luos.h"
 #include "cmd.h"
 #include "convert.h"
@@ -12,4 +11,4 @@ void format_data(container_t *container, char *json);
 unsigned int get_delay(void);
 void set_delay(unsigned int new_delayms);
 
-#endif /* ROUTINGTABLE */
+#endif /* LUOS_TO_JSON */
